+++ packages/core/src/compute-view/deployment-view/__test__/include.deployment.spec.ts
@@ -144,30 +144,29 @@ describe('DeploymentRefPredicate', () => {
       edgeIds: [
         'customer:prod.us',
         'global:customer',
+        "global:dev.devCustomer",
         'global:acc.testCustomer',
         'prod.us:global',
+        "dev.devCloud:global",
         'acc.eu:global',
+        'dev.devCustomer:dev.devCloud',
         'acc.testCustomer:acc.eu',
         'customer:prod.eu.zone1.ui',
         'prod.eu.zone1.ui:prod.eu.zone1.api',
-        'dev.devCustomer:dev.devCloud',
         'prod.eu.zone1.api:global',
         'prod.eu:prod.us',
-        'dev.devCustomer:dev.devCloud',
-        'dev.devCloud:global.email',
-        'global.email:dev.devCustomer',
       ],
       nodeIds: [
         'dev',
-        'dev.devCustomer',
+        'dev.devCloud',
         'acc',
         'acc.eu',
         'global',
         'customer',
-        'dev.devCloud',
         'prod',
         'prod.eu',
         'prod.eu.zone1',
+        'dev.devCustomer',
         'acc.testCustomer',
         'prod.eu.zone1.ui',
         'prod.us',
