+++ packages/core/src/compute-view/deployment-view/__test__/include.relations.spec.ts
@@ -16,7 +16,7 @@ describe('DirectRelationPredicate', () => {
         "Nodes": [
           "global.email",
           "customer.instance",
-          "dev.devCloud.instance",
+          "dev",
           "acc",
           "prod",
           "prod.eu",
@@ -24,7 +24,9 @@ describe('DirectRelationPredicate', () => {
           "prod.eu.zone2",
           "prod.us",
           "prod.us.zone1",
+          "dev.devCustomer.instance",
           "acc.testCustomer.instance",
+          "dev.devCloud.instance",
           "acc.eu",
           "prod.eu.zone1.ui",
           "prod.eu.zone2.ui",
@@ -40,15 +42,12 @@ describe('DirectRelationPredicate', () => {
           "prod.us.zone1.ui",
           "prod.us.zone1.api",
           "prod.us.db",
-          "dev.devCustomer.instance",
-          "acc.testCustomer.instance",
         ],
         "edges": [
           "customer.instance:prod.eu.zone1.ui",
           "customer.instance:prod.eu.zone2.ui",
           "customer.instance:prod.us.zone1.ui",
           "global.email:customer.instance",
-          "dev.devCloud.instance:global.email",
           "prod.eu.zone1.api:prod.eu.auth",
           "prod.eu.zone1.api:prod.eu.media",
           "prod.eu.zone1.api:prod.eu.db",
@@ -56,7 +55,6 @@ describe('DirectRelationPredicate', () => {
           "prod.eu.zone1.ui:prod.eu.zone1.api",
           "prod.eu.zone1.ui:prod.eu.auth",
           "prod.eu.zone1.ui:prod.eu.media",
-          "global.email:dev.devCustomer.instance",
           "prod.eu.zone2.api:prod.eu.auth",
           "prod.eu.zone2.api:prod.eu.media",
           "prod.eu.zone2.api:prod.eu.db",
@@ -68,6 +66,9 @@ describe('DirectRelationPredicate', () => {
           "prod.us.zone1.api:prod.us.db",
           "prod.us.zone1.api:global.email",
           "prod.us.zone1.ui:prod.us.zone1.api",
+          "dev.devCustomer.instance:dev.devCloud.instance",
+          "global.email:dev.devCustomer.instance",
+          "dev.devCloud.instance:global.email",
           "acc.testCustomer.instance:acc.eu.ui",
           "global.email:acc.testCustomer.instance",
           "acc.eu.api:acc.eu.auth",
@@ -75,7 +76,6 @@ describe('DirectRelationPredicate', () => {
           "acc.eu.api:global.email",
           "acc.eu.ui:acc.eu.api",
           "acc.eu.ui:acc.eu.auth",
-          "dev.devCustomer.instance:dev.devCloud.instance",
         ],
       }
     `)
