+++ packages/core/src/compute-view/deployment-view/__test__/relation.incoming.spec.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
-import type { DeploymentViewRuleBuilderOp } from '../../../builder/Builder.view'
-import { $exclude, $include, computeNodesAndEdges, type Types } from './fixture'
+import type { DeploymentRulesBuilderOp } from '../../../builder'
+import { type Types, $exclude, $include, computeNodesAndEdges } from './fixture'
 
-function expectComputed(...rules: DeploymentViewRuleBuilderOp<Types>[]) {
+function expectComputed(...rules: DeploymentRulesBuilderOp<Types>[]) {
   return expect(computeNodesAndEdges(...rules))
 }
 
@@ -10,11 +10,12 @@ describe('IncommingRelationPredicate', () => {
   it('should include relation incoming to deployment node if it matches condition', () => {
     expectComputed(
       $include('customer'),
-      $include('-> prod.eu.zone1.ui', { where: 'tag is #old' })
+      $include('-> prod.eu.zone1.ui', { where: 'tag is #old' }),
     ).toMatchInlineSnapshot(`
       {
         "Nodes": [
           "customer",
+          "prod",
           "prod.eu.zone1.ui",
         ],
         "edges": [
@@ -27,7 +28,7 @@ describe('IncommingRelationPredicate', () => {
   it('should not include relation incoming to deployment node if it does not match condition', () => {
     expectComputed(
       $include('customer'),
-      $include('-> prod.eu.zone1.ui', { where: 'tag is #next' })
+      $include('-> prod.eu.zone1.ui', { where: 'tag is #next' }),
     ).toMatchInlineSnapshot(`
       {
         "Nodes": [
@@ -46,10 +47,10 @@ describe('IncommingRelationPredicate', () => {
       $exclude({
         incoming: {
           ref: {
-            model: 'cloud.backend'
-          }
-        }
-      }, { where: 'tag is #old' })
+            model: 'cloud.backend',
+          },
+        },
+      }, { where: 'tag is #old' }),
     ).toEqual({
       Nodes: [
         'prod.eu.zone1',
@@ -60,12 +61,12 @@ describe('IncommingRelationPredicate', () => {
         'prod.eu.media',
       ],
       edges: [
-        "prod.eu.zone1.api:prod.eu.auth",
-        "prod.eu.zone1.api:prod.eu.media",
-        "prod.eu.zone1.api:prod.eu.db",
-        "prod.eu.zone1.ui:prod.eu.auth",
-        "prod.eu.zone1.ui:prod.eu.media",
-      ]
+        'prod.eu.zone1.api:prod.eu.auth',
+        'prod.eu.zone1.api:prod.eu.media',
+        'prod.eu.zone1.api:prod.eu.db',
+        'prod.eu.zone1.ui:prod.eu.auth',
+        'prod.eu.zone1.ui:prod.eu.media',
+      ],
     })
   })
 
@@ -77,35 +78,36 @@ describe('IncommingRelationPredicate', () => {
       $exclude({
         incoming: {
           ref: {
-            model: 'cloud.backend'
-          }
-        }
-      }, { where: 'tag is #next' })
+            model: 'cloud.backend',
+          },
+        },
+      }, { where: 'tag is #next' }),
     ).toEqual({
       Nodes: [
-        "prod.eu.zone1",
-        "prod.eu.zone1.ui",
-        "prod.eu.zone1.api",
-        "prod.eu.auth",
-        "prod.eu.media",
-        "prod.eu.db",
+        'prod.eu.zone1',
+        'prod.eu.zone1.ui',
+        'prod.eu.zone1.api',
+        'prod.eu.auth',
+        'prod.eu.media',
+        'prod.eu.db',
       ],
       edges: [
-        "prod.eu.zone1.ui:prod.eu.zone1.api",
-        "prod.eu.zone1.api:prod.eu.auth",
-        "prod.eu.zone1.api:prod.eu.media",
-        "prod.eu.zone1.api:prod.eu.db",
-        "prod.eu.zone1.ui:prod.eu.auth",
-        "prod.eu.zone1.ui:prod.eu.media",]
+        'prod.eu.zone1.ui:prod.eu.zone1.api',
+        'prod.eu.zone1.api:prod.eu.auth',
+        'prod.eu.zone1.api:prod.eu.media',
+        'prod.eu.zone1.api:prod.eu.db',
+        'prod.eu.zone1.ui:prod.eu.auth',
+        'prod.eu.zone1.ui:prod.eu.media',
+      ],
     })
-  })  
+  })
 
   it('should exclude relation incoming to deployment node if it matches condition', () => {
     expectComputed(
       $include('prod.eu.*'),
       $include('prod.eu.zone1.**'),
       $exclude('prod.eu.zone2'),
-      $exclude("-> prod.eu.zone1.*", { where: 'tag is #old' })
+      $exclude('-> prod.eu.zone1.*', { where: 'tag is #old' }),
     ).toEqual({
       Nodes: [
         'prod.eu.zone1',
@@ -116,12 +118,12 @@ describe('IncommingRelationPredicate', () => {
         'prod.eu.media',
       ],
       edges: [
-        "prod.eu.zone1.api:prod.eu.auth",
-        "prod.eu.zone1.api:prod.eu.media",
-        "prod.eu.zone1.api:prod.eu.db",
-        "prod.eu.zone1.ui:prod.eu.auth",
-        "prod.eu.zone1.ui:prod.eu.media",
-      ]
+        'prod.eu.zone1.api:prod.eu.auth',
+        'prod.eu.zone1.api:prod.eu.media',
+        'prod.eu.zone1.api:prod.eu.db',
+        'prod.eu.zone1.ui:prod.eu.auth',
+        'prod.eu.zone1.ui:prod.eu.media',
+      ],
     })
   })
 
@@ -130,23 +132,24 @@ describe('IncommingRelationPredicate', () => {
       $include('prod.eu.*'),
       $include('prod.eu.zone1.**'),
       $exclude('prod.eu.zone2'),
-      $exclude("-> prod.eu.zone1.*", { where: 'tag is #next' })
+      $exclude('-> prod.eu.zone1.*', { where: 'tag is #next' }),
     ).toEqual({
       Nodes: [
-        "prod.eu.zone1",
-        "prod.eu.zone1.ui",
-        "prod.eu.zone1.api",
-        "prod.eu.auth",
-        "prod.eu.media",
-        "prod.eu.db",
+        'prod.eu.zone1',
+        'prod.eu.zone1.ui',
+        'prod.eu.zone1.api',
+        'prod.eu.auth',
+        'prod.eu.media',
+        'prod.eu.db',
       ],
       edges: [
-        "prod.eu.zone1.ui:prod.eu.zone1.api",
-        "prod.eu.zone1.api:prod.eu.auth",
-        "prod.eu.zone1.api:prod.eu.media",
-        "prod.eu.zone1.api:prod.eu.db",
-        "prod.eu.zone1.ui:prod.eu.auth",
-        "prod.eu.zone1.ui:prod.eu.media",]
+        'prod.eu.zone1.ui:prod.eu.zone1.api',
+        'prod.eu.zone1.api:prod.eu.auth',
+        'prod.eu.zone1.api:prod.eu.media',
+        'prod.eu.zone1.api:prod.eu.db',
+        'prod.eu.zone1.ui:prod.eu.auth',
+        'prod.eu.zone1.ui:prod.eu.media',
+      ],
     })
   })
 })
