+++ packages/core/src/compute-view/deployment-view/__test__/wildcard.spec.ts
@@ -7,18 +7,15 @@ describe('deployment view: wildcard', () => {
       $include('*'),
     )
     expect(nodeIds).toEqual([
-      'customer.instance',
       'dev',
-      'dev.devCustomer',
+      'dev.devCloud',
       'acc',
       'acc.eu',
       'global',
       'customer',
       'prod',
+      'dev.devCustomer',
       'acc.testCustomer',
-      'prod',
-      'dev.devCloud',
-      'acc.eu',
       'prod.eu',
       'prod.us',
     ])
@@ -26,22 +23,15 @@ describe('deployment view: wildcard', () => {
       'customer:prod.eu',
       'customer:prod.us',
       'global:customer',
+      'global:dev.devCustomer',
       'global:acc.testCustomer',
       'prod.eu:global',
       'prod.us:global',
+      'dev.devCloud:global',
       'acc.eu:global',
       'prod.eu:prod.us',
       'dev.devCustomer:dev.devCloud',
       'acc.testCustomer:acc.eu',
-      'prod.eu:global.email',
-      'prod.us:global.email',
-      'dev.devCloud:global.email',
-      'acc.eu:global.email',
-      'customer.instance:prod.eu',
-      'customer.instance:prod.us',
-      'global.email:dev.devCustomer',
-      'global.email:acc.testCustomer',
-      'global.email:customer.instance',
     ])
   })
